Query Opencast for a specific event

Returns a single event.

oc_event(eventId = "", sign = FALSE, withacl = FALSE,
  withmetadata = FALSE, withscheduling = FALSE,
  withpublications = FALSE)

Arguments

eventId
The event id

sign
Whether public distribution urls should be signed.

withacl
Whether the acl metadata should be included in the response.

withmetadata
Whether the metadata catalogs should be included in the response.

withscheduling
Whether the scheduling information should be included in the response.

withpublications
Whether the publication ids and urls should be included in the response.

Value

Returns a single event.

Details

This function expects the hostname, username and password to be set as environment variables.

Examples

Sys.setenv(OPENCAST_HOST = "https://legacy.opencast.org")
Sys.setenv(OPENCAST_USERNAME = "admin")
Sys.setenv(OPENCAST_PASSWORD = "opencast")

oc_event("91df4c57-9631-4a84-b814-d8e9b36a3c32")

#> $content
#> $content$identifier
#> [1] "91df4c57-9631-4a84-b814-d8e9b36a3c32"
#> 
#> $content$creator
#> [1] "System User"
#> 
#> $content$presenter
#> $content$presenter[[1]]
#> [1] "Durian Open Movie Team"
#> 
#> 
#> $content$created
#> [1] "2019-10-22T00:06:00Z"
#> 
#> $content$is_part_of
#> [1] "446eed07-853b-4865-9813-07e1fb39d2d5"
#> 
#> $content$subjects
#> list()
#> 
#> $content$start
#> [1] "2019-10-22T00:06:16Z"
#> 
#> $content$description
#> [1] "Trailer for the Sintel open movie project"
#> 
#> $content$language
#> [1] ""
#> 
#> $content$source
#> [1] ""
#> 
#> $content$title
#> [1] "Sintel Trailer"
#> 
#> $content$processing_state
#> [1] "SUCCEEDED"
#> 
#> $content$license
#> [1] ""
#> 
#> $content$archive_version
#> [1] 1
#> 
#> $content$contributor
#> list()
#> 
#> $content$series
#> [1] "Blender Foundation Productions"
#> 
#> $content$has_previews
#> [1] TRUE
#> 
#> $content$location
#> [1] ""
#> 
#> $content$rightsholder
#> [1] ""
#> 
#> $content$publication_status
#> $content$publication_status[[1]]
#> [1] "internal"
#> 
#> $content$publication_status[[2]]
#> [1] "engage-player"
#> 
#> $content$publication_status[[3]]
#> [1] "oaipmh-default"
#> 
#> 
#> $content$status
#> [1] "EVENTS.EVENTS.STATUS.PROCESSED"
#> 
#> 
#> $url
#> [1] "https://legacy.opencast.org/api/events/91df4c57-9631-4a84-b814-d8e9b36a3c32"
#> 
#> $response
#> Response [https://legacy.opencast.org/api/events/91df4c57-9631-4a84-b814-d8e9b36a3c32?sign=FALSE&withacl=FALSE&withmetadata=FALSE&withscheduling=FALSE]
#>   Date: 2019-10-22 18:31
#>   Status: 200
#>   Content-Type: application/v1.2.0+json
#>   Size: 639 B
#> <BINARY BODY>
#> 
#> attr(,"class")
#> [1] "opencast_api"