Query Opencast for the list of workflow instances

Returns a list of workflow instances.

oc_list_workflow_instances(withoperations = FALSE,
  withconfigurationpanel = FALSE, filter = "", sort = "",
  limit = 0, offset = 0)

Arguments

withoperations
Whether the workflow operations should be included in the response.

withconfigurationpanel
Whether the workflow configuration panel should be included in the response.

filter
A comma seperated list of filters to limit the results with. A filter is the filter's name followed by a colon ":" and then the value to filter with so it is the form :.

sort
Sort the results based upon a list of comma seperated sorting criteria. In the comma seperated list each type of sorting is specified as a pair such as: :ASC or :DESC. Adding the suffix ASC or DESC sets the order as ascending or descending order and is mandatory.

limit
The maximum number of results to return for a single request.

offset
The index of the first result to return.

Value

Returns a list of workflow instances.

Details

This function expects the hostname, username and password to be set as environment variables.

Examples

Sys.setenv(OPENCAST_HOST = "https://legacy.opencast.org")
Sys.setenv(OPENCAST_USERNAME = "admin")
Sys.setenv(OPENCAST_PASSWORD = "opencast")

oc_list_workflow_instances()

#> $content
#>   workflow_definition_identifier identifier                        creator
#> 1                           fast        524                    System User
#> 2                           fast        521                    System User
#> 3                           fast        533                    System User
#> 4                        retract        896 Opencast Project Administrator
#> 5                           fast        546                    System User
#> 6                           fast        543                    System User
#>                                                                                                                                                                                                    description
#> 1 \n    A minimal workflow that transcodes the media into distribution formats, then\n    sends the resulting distribution files, along with their associated metadata,\n    to the distribution channels.\n  
#> 2 \n    A minimal workflow that transcodes the media into distribution formats, then\n    sends the resulting distribution files, along with their associated metadata,\n    to the distribution channels.\n  
#> 3 \n    A minimal workflow that transcodes the media into distribution formats, then\n    sends the resulting distribution files, along with their associated metadata,\n    to the distribution channels.\n  
#> 4                                                                                                                                                                                          Retract a recording
#> 5 \n    A minimal workflow that transcodes the media into distribution formats, then\n    sends the resulting distribution files, along with their associated metadata,\n    to the distribution channels.\n  
#> 6 \n    A minimal workflow that transcodes the media into distribution formats, then\n    sends the resulting distribution files, along with their associated metadata,\n    to the distribution channels.\n  
#>       state                 title                     event_identifier
#> 1 succeeded Fast Testing Workflow ab92cc10-c422-4d54-99db-055563658897
#> 2 succeeded Fast Testing Workflow cfa2d07e-389a-4785-aa4d-49e0c2073f13
#> 3 succeeded Fast Testing Workflow 869a0e15-4a65-4363-8f40-971b40e6bc05
#> 4 succeeded               Retract cfa2d07e-389a-4785-aa4d-49e0c2073f13
#> 5 succeeded Fast Testing Workflow 91708b2e-285e-4f08-855d-2ad882c42002
#> 6 succeeded Fast Testing Workflow a05cd45a-606c-4076-a60f-44f1cf0a38f1
#> 
#> $url
#> [1] "https://legacy.opencast.org/api/workflows/"
#> 
#> $response
#> Response [https://legacy.opencast.org/api/workflows/?withoperations=FALSE&withconfigurationpanel=FALSE&filter=&sort=&limit=0&offset=0]
#>   Date: 2019-11-04 11:18
#>   Status: 200
#>   Content-Type: application/v1.2.0+json
#>   Size: 2.31 kB
#> <BINARY BODY>
#> 
#> attr(,"class")
#> [1] "opencast_api"