Print result of oc_events()

Print a structured return of the oc_events() function.

# S3 method for oc_events
print(x, ...)

Arguments

x
The return of the function this print function relates to.

...
Possible further options to the print function.

Value

A structured print of the return by the oc_events() function.

See also

Examples

Sys.setenv(OPENCAST_HOST = "https://legacy.opencast.org")
Sys.setenv(OPENCAST_USERNAME = "admin")
Sys.setenv(OPENCAST_PASSWORD = "opencast")

resp <- oc_events()

resp

#> $content
#> $content[[1]]
#> $content[[1]]$identifier
#> [1] "91df4c57-9631-4a84-b814-d8e9b36a3c32"
#> 
#> $content[[1]]$creator
#> [1] "System User"
#> 
#> $content[[1]]$presenter
#> $content[[1]]$presenter[[1]]
#> [1] "Durian Open Movie Team"
#> 
#> 
#> $content[[1]]$created
#> [1] "2019-10-22T00:06:00Z"
#> 
#> $content[[1]]$is_part_of
#> [1] "446eed07-853b-4865-9813-07e1fb39d2d5"
#> 
#> $content[[1]]$subjects
#> list()
#> 
#> $content[[1]]$start
#> [1] "2019-10-22T00:06:16Z"
#> 
#> $content[[1]]$description
#> [1] "Trailer for the Sintel open movie project"
#> 
#> $content[[1]]$language
#> [1] ""
#> 
#> $content[[1]]$source
#> [1] ""
#> 
#> $content[[1]]$title
#> [1] "Sintel Trailer"
#> 
#> $content[[1]]$processing_state
#> [1] "SUCCEEDED"
#> 
#> $content[[1]]$license
#> [1] ""
#> 
#> $content[[1]]$archive_version
#> [1] 1
#> 
#> $content[[1]]$contributor
#> list()
#> 
#> $content[[1]]$series
#> [1] "Blender Foundation Productions"
#> 
#> $content[[1]]$has_previews
#> [1] TRUE
#> 
#> $content[[1]]$location
#> [1] ""
#> 
#> $content[[1]]$rightsholder
#> [1] ""
#> 
#> $content[[1]]$publication_status
#> $content[[1]]$publication_status[[1]]
#> [1] "internal"
#> 
#> $content[[1]]$publication_status[[2]]
#> [1] "engage-player"
#> 
#> $content[[1]]$publication_status[[3]]
#> [1] "oaipmh-default"
#> 
#> 
#> $content[[1]]$status
#> [1] "EVENTS.EVENTS.STATUS.PROCESSED"
#> 
#> 
#> $content[[2]]
#> $content[[2]]$identifier
#> [1] "b31ab134-c460-45fa-a40a-b475ca3bc2ce"
#> 
#> $content[[2]]$creator
#> [1] "System User"
#> 
#> $content[[2]]$presenter
#> $content[[2]]$presenter[[1]]
#> [1] "Lars Kiesow"
#> 
#> 
#> $content[[2]]$created
#> [1] "2019-10-22T00:05:00Z"
#> 
#> $content[[2]]$is_part_of
#> [1] ""
#> 
#> $content[[2]]$subjects
#> list()
#> 
#> $content[[2]]$start
#> [1] "2019-10-22T00:05:50Z"
#> 
#> $content[[2]]$description
#> [1] "Opencast quick Jira ticket and pull request creator, https://github.com/lkiesow/ocpr"
#> 
#> $content[[2]]$language
#> [1] ""
#> 
#> $content[[2]]$source
#> [1] ""
#> 
#> $content[[2]]$title
#> [1] "OCPR Demo"
#> 
#> $content[[2]]$processing_state
#> [1] "SUCCEEDED"
#> 
#> $content[[2]]$license
#> [1] ""
#> 
#> $content[[2]]$archive_version
#> [1] 1
#> 
#> $content[[2]]$contributor
#> list()
#> 
#> $content[[2]]$series
#> [1] ""
#> 
#> $content[[2]]$has_previews
#> [1] TRUE
#> 
#> $content[[2]]$location
#> [1] ""
#> 
#> $content[[2]]$rightsholder
#> [1] ""
#> 
#> $content[[2]]$publication_status
#> $content[[2]]$publication_status[[1]]
#> [1] "internal"
#> 
#> $content[[2]]$publication_status[[2]]
#> [1] "engage-player"
#> 
#> $content[[2]]$publication_status[[3]]
#> [1] "oaipmh-default"
#> 
#> 
#> $content[[2]]$status
#> [1] "EVENTS.EVENTS.STATUS.PROCESSED"
#> 
#> 
#> $content[[3]]
#> $content[[3]]$identifier
#> [1] "97ba18eb-40de-4796-a9e8-74945974a42b"
#> 
#> $content[[3]]$creator
#> [1] "System User"
#> 
#> $content[[3]]$presenter
#> $content[[3]]$presenter[[1]]
#> [1] "Olaf Schulte"
#> 
#> 
#> $content[[3]]$created
#> [1] "2019-10-22T00:05:00Z"
#> 
#> $content[[3]]$is_part_of
#> [1] ""
#> 
#> $content[[3]]$subjects
#> list()
#> 
#> $content[[3]]$start
#> [1] "2019-10-22T00:05:49Z"
#> 
#> $content[[3]]$description
#> [1] ""
#> 
#> $content[[3]]$language
#> [1] ""
#> 
#> $content[[3]]$source
#> [1] ""
#> 
#> $content[[3]]$title
#> [1] "About Opencast"
#> 
#> $content[[3]]$processing_state
#> [1] "SUCCEEDED"
#> 
#> $content[[3]]$license
#> [1] ""
#> 
#> $content[[3]]$archive_version
#> [1] 1
#> 
#> $content[[3]]$contributor
#> list()
#> 
#> $content[[3]]$series
#> [1] ""
#> 
#> $content[[3]]$has_previews
#> [1] TRUE
#> 
#> $content[[3]]$location
#> [1] ""
#> 
#> $content[[3]]$rightsholder
#> [1] ""
#> 
#> $content[[3]]$publication_status
#> $content[[3]]$publication_status[[1]]
#> [1] "internal"
#> 
#> $content[[3]]$publication_status[[2]]
#> [1] "engage-player"
#> 
#> $content[[3]]$publication_status[[3]]
#> [1] "oaipmh-default"
#> 
#> 
#> $content[[3]]$status
#> [1] "EVENTS.EVENTS.STATUS.PROCESSED"
#> 
#> 
#> $content[[4]]
#> $content[[4]]$identifier
#> [1] "03633d4e-4742-4fd8-8b39-e03783f9839f"
#> 
#> $content[[4]]$creator
#> [1] "System User"
#> 
#> $content[[4]]$presenter
#> $content[[4]]$presenter[[1]]
#> [1] "Blender Foundation"
#> 
#> 
#> $content[[4]]$created
#> [1] "2019-10-22T00:05:00Z"
#> 
#> $content[[4]]$is_part_of
#> [1] "446eed07-853b-4865-9813-07e1fb39d2d5"
#> 
#> $content[[4]]$subjects
#> list()
#> 
#> $content[[4]]$start
#> [1] "2019-10-22T00:06:02Z"
#> 
#> $content[[4]]$description
#> [1] ""
#> 
#> $content[[4]]$language
#> [1] ""
#> 
#> $content[[4]]$source
#> [1] ""
#> 
#> $content[[4]]$title
#> [1] "Tears of Steel"
#> 
#> $content[[4]]$processing_state
#> [1] "SUCCEEDED"
#> 
#> $content[[4]]$license
#> [1] ""
#> 
#> $content[[4]]$archive_version
#> [1] 1
#> 
#> $content[[4]]$contributor
#> list()
#> 
#> $content[[4]]$series
#> [1] "Blender Foundation Productions"
#> 
#> $content[[4]]$has_previews
#> [1] TRUE
#> 
#> $content[[4]]$location
#> [1] ""
#> 
#> $content[[4]]$rightsholder
#> [1] ""
#> 
#> $content[[4]]$publication_status
#> $content[[4]]$publication_status[[1]]
#> [1] "internal"
#> 
#> $content[[4]]$publication_status[[2]]
#> [1] "engage-player"
#> 
#> $content[[4]]$publication_status[[3]]
#> [1] "oaipmh-default"
#> 
#> 
#> $content[[4]]$status
#> [1] "EVENTS.EVENTS.STATUS.PROCESSED"
#> 
#> 
#> $content[[5]]
#> $content[[5]]$identifier
#> [1] "2c76606a-000d-4d23-9a56-7a777947ae21"
#> 
#> $content[[5]]$creator
#> [1] "System User"
#> 
#> $content[[5]]$presenter
#> $content[[5]]$presenter[[1]]
#> [1] "Blender Foundation"
#> 
#> 
#> $content[[5]]$created
#> [1] "2019-10-22T00:06:00Z"
#> 
#> $content[[5]]$is_part_of
#> [1] "446eed07-853b-4865-9813-07e1fb39d2d5"
#> 
#> $content[[5]]$subjects
#> list()
#> 
#> $content[[5]]$start
#> [1] "2019-10-22T00:06:16Z"
#> 
#> $content[[5]]$description
#> [1] ""
#> 
#> $content[[5]]$language
#> [1] ""
#> 
#> $content[[5]]$source
#> [1] ""
#> 
#> $content[[5]]$title
#> [1] "Tears of Steel (Dualstream)"
#> 
#> $content[[5]]$processing_state
#> [1] "SUCCEEDED"
#> 
#> $content[[5]]$license
#> [1] ""
#> 
#> $content[[5]]$archive_version
#> [1] 1
#> 
#> $content[[5]]$contributor
#> list()
#> 
#> $content[[5]]$series
#> [1] "Blender Foundation Productions"
#> 
#> $content[[5]]$has_previews
#> [1] TRUE
#> 
#> $content[[5]]$location
#> [1] ""
#> 
#> $content[[5]]$rightsholder
#> [1] ""
#> 
#> $content[[5]]$publication_status
#> $content[[5]]$publication_status[[1]]
#> [1] "internal"
#> 
#> $content[[5]]$publication_status[[2]]
#> [1] "engage-player"
#> 
#> $content[[5]]$publication_status[[3]]
#> [1] "oaipmh-default"
#> 
#> 
#> $content[[5]]$status
#> [1] "EVENTS.EVENTS.STATUS.PROCESSED"
#> 
#> 
#> 
#> $url
#> [1] "https://legacy.opencast.org/api/events/"
#> 
#> $response
#> Response [https://legacy.opencast.org/api/events/?sign=FALSE&withacl=FALSE&withmetadata=FALSE&withscheduling=FALSE&filter=&sort=&limit=0&offset=0]
#>   Date: 2019-10-22 18:42
#>   Status: 200
#>   Content-Type: application/v1.2.0+json
#>   Size: 2.97 kB
#> <BINARY BODY>
#> 
#> attr(,"class")
#> [1] "opencast_api"