Print result of oc_list_events()

Print a structured return of the oc_list_events() function.

# S3 method for oc_list_events
print(x, ...)

Arguments

x
The return of the function this print function relates to.

...
Possible further options to the print function.

Value

A structured print of the return by the oc_list_events() function.

See also

Examples

Sys.setenv(OPENCAST_HOST = "https://legacy.opencast.org")
Sys.setenv(OPENCAST_USERNAME = "admin")
Sys.setenv(OPENCAST_PASSWORD = "opencast")

resp <- oc_list_events()

resp

#> $content
#>                             identifier                        creator
#> 1 71061353-56cc-4249-8b8c-6ae7e4fa694d Opencast Project Administrator
#> 2 2c028936-6610-4ea2-af4e-bc88eb6ee50a Opencast Project Administrator
#> 3 4bfb829b-fe30-4ca4-8387-b5921781e6af Opencast Project Administrator
#> 4 bb69d37c-f842-4818-acf5-1e7a6986e2c1 Opencast Project Administrator
#> 5 5f7b4557-e6c3-4669-a0f1-e7484d66617a Opencast Project Administrator
#>                presenter              created
#> 1 Durian Open Movie Team 2020-03-19T01:04:00Z
#> 2            Lars Kiesow 2020-03-19T01:03:00Z
#> 3           Olaf Schulte 2020-03-19T01:03:00Z
#> 4     Blender Foundation 2020-03-19T01:03:00Z
#> 5     Blender Foundation 2020-03-19T01:03:00Z
#>                             is_part_of subjects                start
#> 1 706543e3-c731-41de-b6c7-ed28c90b3635     NULL 2020-03-19T01:04:06Z
#> 2                                          NULL 2020-03-19T01:03:39Z
#> 3                                          NULL 2020-03-19T01:03:39Z
#> 4 706543e3-c731-41de-b6c7-ed28c90b3635     NULL 2020-03-19T01:03:52Z
#> 5 706543e3-c731-41de-b6c7-ed28c90b3635     NULL 2020-03-19T01:04:06Z
#>                                                                            description
#> 1                                            Trailer for the Sintel open movie project
#> 2 Opencast quick Jira ticket and pull request creator, https://github.com/lkiesow/ocpr
#> 3                                                                                     
#> 4                                                                                     
#> 5                                                                                     
#>   language source                       title processing_state license
#> 1                              Sintel Trailer        SUCCEEDED        
#> 2                                   OCPR Demo        SUCCEEDED        
#> 3                              About Opencast        SUCCEEDED        
#> 4                              Tears of Steel        SUCCEEDED        
#> 5                 Tears of Steel (Dualstream)        SUCCEEDED        
#>   archive_version contributor                         series has_previews
#> 1               1        NULL Blender Foundation Productions         TRUE
#> 2               1        NULL                                        TRUE
#> 3               1        NULL                                        TRUE
#> 4               1        NULL Blender Foundation Productions         TRUE
#> 5               1        NULL Blender Foundation Productions         TRUE
#>   location rightsholder                      publication_status
#> 1                       internal, engage-player, oaipmh-default
#> 2                       internal, engage-player, oaipmh-default
#> 3                       internal, engage-player, oaipmh-default
#> 4                       internal, engage-player, oaipmh-default
#> 5                       internal, engage-player, oaipmh-default
#>                           status
#> 1 EVENTS.EVENTS.STATUS.PROCESSED
#> 2 EVENTS.EVENTS.STATUS.PROCESSED
#> 3 EVENTS.EVENTS.STATUS.PROCESSED
#> 4 EVENTS.EVENTS.STATUS.PROCESSED
#> 5 EVENTS.EVENTS.STATUS.PROCESSED
#> 
#> $url
#> [1] "https://legacy.opencast.org/api/events/"
#> 
#> $response
#> Response [https://legacy.opencast.org/api/events/?sign=FALSE&withacl=FALSE&withmetadata=FALSE&withscheduling=FALSE&filter=&sort=&limit=0&offset=0]
#>   Date: 2020-03-19 11:21
#>   Status: 200
#>   Content-Type: application/v1.2.0+json
#>   Size: 3.07 kB
#> <BINARY BODY>
#> 
#> attr(,"class")
#> [1] "opencast_api"