Print result of oc_list_workflow_definitions()

Print a structured return of the oc_list_workflow_definitions() function.

# S3 method for oc_list_workflow_definitions
print(x, ...)

Arguments

x
The return of the function this print function relates to.

...
Possible further options to the print function.

Value

A structured print of the return by the oc_list_workflow_definitions() function.

See also

Examples

Sys.setenv(OPENCAST_HOST = "https://legacy.opencast.org")
Sys.setenv(OPENCAST_USERNAME = "admin")
Sys.setenv(OPENCAST_PASSWORD = "opencast")

resp <- oc_list_workflow_definitions()

resp

#> $content
#>                   identifier
#> 1              partial-error
#> 2            partial-cleanup
#> 3  attach-watson-transcripts
#> 4           placeholdervideo
#> 5            duplicate-event
#> 6                       fast
#> 7                     import
#> 8             partial-ingest
#> 9          offload-to-remote
#> 10           partial-preview
#> 11       partial-title-slide
#> 12              partial-work
#> 13       schedule-and-upload
#> 14                   publish
#> 15     publish-after-cutting
#> 16           partial-publish
#> 17   publish-uploaded-assets
#> 18         partial-watermark
#> 19        republish-metadata
#> 20       restore-from-remote
#> 21                   retract
#> 22  retry-watson-transcripts
#> 23           update-previews
#>                                                                                                                                                                                                     description
#> 1                                                                                                                                                                            Cleanup after a processing failure
#> 2                                                                                                                                                                                      Cleanup after processing
#> 3                                                         Attach automated transcription generated by the IBM\n    Watson service. This is an internal workflow, started by the\n    Transcription Service.\n  
#> 4                                                                    \n    Publish a placeholder video for corrupt videos.\n    Delete and cleanup corrupt videos and old snapshots, but keep the metadata.\n  
#> 5                                                                                                                                                                                                              
#> 6  \n    A minimal workflow that transcodes the media into distribution formats, then\n    sends the resulting distribution files, along with their associated metadata,\n    to the distribution channels.\n  
#> 7                                                                                                                                                 \n  A minimal Workflow that only publishes a Mediapackage\n  
#> 8                                                                                                                                                                                                              
#> 9                                                                                                                                                                                                              
#> 10                                                                                                                                                                                                             
#> 11                                                                                                                                            \n     Create a title slide video based on the trimmed tracks\n  
#> 12                                                                                                                                                                                                             
#> 13                                                                                                                                        \n    A complex workflow, showcasing most of Opencast's features.\n  
#> 14                                                                                                                                                                                                             
#> 15                                                                                                                                                                                                             
#> 16                                                                                                                                                                                                             
#> 17                                                                                                                                                                                      Publish uploaded assets
#> 18                                                                                                                                           \n     Render watermark into presenter and presentation tracks\n  
#> 19                                                                                                                                                                                                             
#> 20                                                                                                                                                                                                             
#> 21                                                                                                                                                                                          Retract a recording
#> 22                                                                                                                                       Retry transcription job in error. Called by the transcription service.
#> 23                                                                                                                                                                                                             
#>                                                 title             tags
#> 1                                                                     
#> 2                                                                     
#> 3  Attach caption/transcripts generated by IBM Watson                 
#> 4               Cleanup and publish placeholder video  archive, editor
#> 5                                     Duplicate Event          archive
#> 6                               Fast Testing Workflow schedule, upload
#> 7    Import Mediapackage and Publish without reencode                 
#> 8                                        Ingest asset                 
#> 9                           Offload to Remote Storage          archive
#> 10                          Prepare preview artifacts                 
#> 11                          Prepare title slide video                 
#> 12                              Prepare work versions                 
#> 13                   Process upon upload and schedule schedule, upload
#> 14                                            Publish          archive
#> 15                                            Publish           editor
#> 16                              Publish the recording                 
#> 17                            Publish uploaded assets                 
#> 18                                   Render watermark                 
#> 19                                 Republish metadata          archive
#> 20                        Restore from Remote Storage          archive
#> 21                                            Retract  archive, delete
#> 22 Start automated caption/transcription job (watson)                 
#> 23                                    Update previews          archive
#> 
#> $url
#> [1] "https://legacy.opencast.org/api/workflow-definitions/"
#> 
#> $response
#> Response [https://legacy.opencast.org/api/workflow-definitions/?withoperations=FALSE&withconfigurationpanel=FALSE&filter=&sort=&limit=0&offset=0]
#>   Date: 2019-11-04 11:19
#>   Status: 200
#>   Content-Type: application/v1.2.0+json
#>   Size: 3.21 kB
#> <BINARY BODY>
#> 
#> attr(,"class")
#> [1] "opencast_api"