Print result of oc_list_events()

Print a structured return of the oc_list_events() function.

# S3 method for oc_list_events
print(x, ...)

Arguments

x
The return of the function this print function relates to.

...
Possible further options to the print function.

Value

A structured print of the return by the oc_list_events() function.

See also

Examples

Sys.setenv(OPENCAST_HOST = "https://legacy.opencast.org")
Sys.setenv(OPENCAST_USERNAME = "admin")
Sys.setenv(OPENCAST_PASSWORD = "opencast")

resp <- oc_list_events()

resp

#> $content
#>                             identifier                        creator
#> 1 91708b2e-285e-4f08-855d-2ad882c42002                    System User
#> 2 ab92cc10-c422-4d54-99db-055563658897                    System User
#> 3 869a0e15-4a65-4363-8f40-971b40e6bc05                    System User
#> 4 a05cd45a-606c-4076-a60f-44f1cf0a38f1                    System User
#> 5 cfa2d07e-389a-4785-aa4d-49e0c2073f13 Opencast Project Administrator
#>                presenter              created
#> 1 Durian Open Movie Team 2019-11-04T01:06:00Z
#> 2            Lars Kiesow 2019-11-04T01:05:00Z
#> 3     Blender Foundation 2019-11-04T01:05:00Z
#> 4     Blender Foundation 2019-11-04T01:06:00Z
#> 5           Olaf Schulte 2019-11-04T01:05:00Z
#>                             is_part_of subjects                start
#> 1 a1b6787e-1cfa-49e7-85ce-09d36677b92e     NULL 2019-11-04T01:06:19Z
#> 2                                          NULL 2019-11-04T01:05:53Z
#> 3 a1b6787e-1cfa-49e7-85ce-09d36677b92e     NULL 2019-11-04T01:06:06Z
#> 4 a1b6787e-1cfa-49e7-85ce-09d36677b92e     NULL 2019-11-04T01:06:19Z
#> 5                                          NULL 2019-11-04T01:05:52Z
#>                                                                            description
#> 1                                            Trailer for the Sintel open movie project
#> 2 Opencast quick Jira ticket and pull request creator, https://github.com/lkiesow/ocpr
#> 3                                                                                     
#> 4                                                                                     
#> 5                                                                                     
#>   language source                       title processing_state license
#> 1                              Sintel Trailer        SUCCEEDED        
#> 2                                   OCPR Demo        SUCCEEDED        
#> 3                              Tears of Steel        SUCCEEDED        
#> 4                 Tears of Steel (Dualstream)        SUCCEEDED        
#> 5                              About Opencast        SUCCEEDED        
#>   archive_version contributor                         series has_previews
#> 1               1        NULL Blender Foundation Productions         TRUE
#> 2               1        NULL                                        TRUE
#> 3               1        NULL Blender Foundation Productions         TRUE
#> 4               1        NULL Blender Foundation Productions         TRUE
#> 5               2        NULL                                       FALSE
#>   location rightsholder                      publication_status
#> 1                       internal, engage-player, oaipmh-default
#> 2                       internal, engage-player, oaipmh-default
#> 3                       internal, engage-player, oaipmh-default
#> 4                       internal, engage-player, oaipmh-default
#> 5                                                              
#>                           status
#> 1 EVENTS.EVENTS.STATUS.PROCESSED
#> 2 EVENTS.EVENTS.STATUS.PROCESSED
#> 3 EVENTS.EVENTS.STATUS.PROCESSED
#> 4 EVENTS.EVENTS.STATUS.PROCESSED
#> 5 EVENTS.EVENTS.STATUS.PROCESSED
#> 
#> $url
#> [1] "https://legacy.opencast.org/api/events/"
#> 
#> $response
#> Response [https://legacy.opencast.org/api/events/?sign=FALSE&withacl=FALSE&withmetadata=FALSE&withscheduling=FALSE&filter=&sort=&limit=0&offset=0]
#>   Date: 2019-11-04 11:19
#>   Status: 200
#>   Content-Type: application/v1.2.0+json
#>   Size: 2.95 kB
#> <BINARY BODY>
#> 
#> attr(,"class")
#> [1] "opencast_api"