Print result of oc_list_events()

Print a structured return of the oc_list_events() function.

# S3 method for oc_list_events
print(x, ...)

Arguments

x
The return of the function this print function relates to.

...
Possible further options to the print function.

Value

A structured print of the return by the oc_list_events() function.

See also

Examples

Sys.setenv(OPENCAST_HOST = "https://legacy.opencast.org")
Sys.setenv(OPENCAST_USERNAME = "admin")
Sys.setenv(OPENCAST_PASSWORD = "opencast")

resp <- oc_list_events()

resp

#> $content
#>                             identifier     creator              presenter
#> 1 dbb56a4e-8f97-48f7-9ae4-bbaa134f5137 System User Durian Open Movie Team
#> 2 d5d8a9cb-1650-476f-883b-cc0511fc5597 System User            Lars Kiesow
#> 3 0d2ca3e0-42f2-43ea-b716-e16dfacff3c8 System User           Olaf Schulte
#> 4 9184505e-dc82-44b2-9200-11b94adff9ca System User     Blender Foundation
#> 5 58384f85-bf1c-46cc-a96b-e3e4963d7bc9 System User     Blender Foundation
#>                created                           is_part_of subjects
#> 1 2019-10-29T01:06:00Z 251e807b-7d0f-4d71-a1a8-1e28281109f3     NULL
#> 2 2019-10-29T01:05:00Z                                          NULL
#> 3 2019-10-29T01:05:00Z                                          NULL
#> 4 2019-10-29T01:05:00Z 251e807b-7d0f-4d71-a1a8-1e28281109f3     NULL
#> 5 2019-10-29T01:06:00Z 251e807b-7d0f-4d71-a1a8-1e28281109f3     NULL
#>                  start
#> 1 2019-10-29T01:06:21Z
#> 2 2019-10-29T01:05:54Z
#> 3 2019-10-29T01:05:54Z
#> 4 2019-10-29T01:06:07Z
#> 5 2019-10-29T01:06:20Z
#>                                                                            description
#> 1                                            Trailer for the Sintel open movie project
#> 2 Opencast quick Jira ticket and pull request creator, https://github.com/lkiesow/ocpr
#> 3                                                                                     
#> 4                                                                                     
#> 5                                                                                     
#>   language source                       title processing_state license
#> 1                              Sintel Trailer        SUCCEEDED        
#> 2                                   OCPR Demo        SUCCEEDED        
#> 3                              About Opencast        SUCCEEDED        
#> 4                              Tears of Steel        SUCCEEDED        
#> 5                 Tears of Steel (Dualstream)        SUCCEEDED        
#>   archive_version contributor                         series has_previews
#> 1               1        NULL Blender Foundation Productions         TRUE
#> 2               1        NULL                                        TRUE
#> 3               1        NULL                                        TRUE
#> 4               1        NULL Blender Foundation Productions         TRUE
#> 5               1        NULL Blender Foundation Productions         TRUE
#>   location rightsholder                      publication_status
#> 1                       internal, engage-player, oaipmh-default
#> 2                       internal, engage-player, oaipmh-default
#> 3                       internal, engage-player, oaipmh-default
#> 4                       internal, engage-player, oaipmh-default
#> 5                       internal, engage-player, oaipmh-default
#>                           status
#> 1 EVENTS.EVENTS.STATUS.PROCESSED
#> 2 EVENTS.EVENTS.STATUS.PROCESSED
#> 3 EVENTS.EVENTS.STATUS.PROCESSED
#> 4 EVENTS.EVENTS.STATUS.PROCESSED
#> 5 EVENTS.EVENTS.STATUS.PROCESSED
#> 
#> $url
#> [1] "https://legacy.opencast.org/api/events/"
#> 
#> $response
#> Response [https://legacy.opencast.org/api/events/?sign=FALSE&withacl=FALSE&withmetadata=FALSE&withscheduling=FALSE&filter=&sort=&limit=0&offset=0]
#>   Date: 2019-10-29 15:08
#>   Status: 200
#>   Content-Type: application/v1.2.0+json
#>   Size: 2.97 kB
#> <BINARY BODY>
#> 
#> attr(,"class")
#> [1] "opencast_api"